import * from django.conf.urls.defaults for v1.2.x support
In django 1.2.x the urlresolver looked at the client url conf
for the handler400/handler500.  If you do not import *
from django.conf.urls.defaults django will fail to find the
necessary handers.

In django 1.3 the urlresolver will first look at the client url conf for the
necessary handers and if they are not found it will
import the default handlers from django.conf.urls