Use getattr to check is GuardedModelAdminMixin has the attribute queryset
We can not check for queryset in Django 1.8 as the attribute is no longer available
The getattr function still checks the default parameter and fails with a AttributeError
as super(GuardedModelAdminMixin, self) has no attribute getquery. We also need to
pass a default parameter (None) to the second getattr call as this will otherwise fail
as well.