Add support for "direct_perms_only" option
This option identifies the permissions which are *directly* assigned
to a user for an object. It disregards permissions that the user may
have acquired via group memberships or superuser status.

The new option is disabled by default for backwards compatibility.

Two new unit tests are added to demonstrate behavior with and without
the option.

The purpose of the feature is to allow proper initialization of permission
forms used to grant specific permissions to specific users,
independently of their superuser or group memberships.