Django 1.8 renamed model attr to related_model
From Django 1.8 release notes: "In earlier versions of Django, on a
model with a reverse foreign key relationship (for example),
model._meta.get_all_related_objects() returned the relationship as a
django.db.models.related.RelatedObject with the model attribute set to
the source of the relationship. Now, this method returns the
relationship as django.db.models.fields.related.ManyToOneRel (private
API RelatedObject has been removed), and the model attribute is set to
the target of the relationship instead of the source. The source model
is accessible on the related_model attribute instead."