Remove friendly_load tag
Was only used to load adminmedia and admin_static libraries.

For adminmedia, loads the  admin_media_prefix tag which we don't use
anywhere. Was also broken for Django 1.9 - didn't ignore load errors
like it was suppose to.

For admin_static, loads the static tag, which is needed, so admin_static
is not optional.