Fix tests
"./manage.py test" puts the working directory in the PYTHONPATH - this
breaks "./setup.py test".

"./setup.py test" installs guardian into a temp directory and adds that
directory to the PYTHONPATH, however Python 3.x gets confused when it
finds the copy from the working directory first.

So we create a temporary directory and change into that first before
running the tests.