Fix up processing of ANONYMOUS_USER_NAME
We need to be able to handle three cases:

1. ANONYMOUS_USER_NAME is specified. Then we create and fetch the
anonymous user by user.USERNAME_FIELD == ANONYMOUS_USER_NAME

If the anonymous user cannot be found, the code to create it can be
overriden by setting GET_INIT_ANONYMOUS_USER.

2. ANONYMOUS_USER_NAME is not specified. Then we default to the value of
"AnonymousUser". We continue as for case 1.

3. ANONYMOUS_USER_NAME is None. In which case the use of the anonymous
user is disabled.

Before this patch, we did not handle case 3 properly. This patch fixes
that.

I believe this maintains API compatability with previous versions of
Guardian before removing ANONYMOUS_USER_ID in
286ced72d7fa925f68ce8b49b3a2adc6cbe6aa6c

Closes #409.