Use the correct DB in post_migration handler
Django provides a keyword argument called `using` to the `post_migrate`
signal handler. This argument tells the handler which database was
migrated.

Previously, the `create_anonymous_user()` function would always try to
use the default database to create the anonymous user. This change uses
the `using` parameter to make sure we use the database that actually
underwent the migration.

Closes: django-guardian/django-guardian#630