create get_objects_for_user_kwargs method on PermissionListMixin and only show this method in the docs
the existing PermissionListMixin had a method called 'get_get_objects_for_user_kwargs' with two 'get_'s. The new method remove the second 'get_' in the function signature but with the same implementation'. We leave the existing method 'get_get_...' available but do not document it as part fo the public API