add support for non-User models in `get_identity` util
inspiration: in multi-tenant applications we require a strict separation
between authentication and authorization

with `get_user_obj_perms_model` we are allowed to set the permission
model that will be used for persisting permissions in the database.
this allows us to point the user FK to a non-User model, however, the
`get_identity` still relies on the User model explicitly.

the only requirement for the permission model is to extend the `PermissionsMixin`
from django, rather to be the `get_user_model`.